pyproject.toml
[project]
name = "game-toolbox"
version = "0.1.0"
description = "Modular media & game-dev toolbox with GUI and pipeline support"
readme = "README.md"
license = { text = "GPL-3.0-only" }
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "click>=8.1",
    "opencv-python-headless>=4.9",
    "Pillow>=10.0",
    "numpy>=2.4.2",
]

[project.optional-dependencies]
audio = ["pydub>=0.25"]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.3",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
game-toolbox = "game_toolbox.cli.main:cli"

[project.gui-scripts]
game-toolbox-gui = "game_toolbox.gui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/game_toolbox"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]
exclude = ["**/extract_frames_legacy.py"]

[tool.ruff.lint]
select = [
    "E", "W",
    "F",
    "I",
    "N",
    "UP",
    "D",
    "B",
    "SIM",
    "RUF",
]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"
packages = ["game_toolbox"]
exclude = ["extract_frames_legacy\\.py$"]

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
pythonpath = ["src"]
